+++ helpers/env.go
@@ -16,6 +16,8 @@ func LoadEnv() {
 	}
 }
 
+// ReadEnvDefault reads an environment variable with a default value if it doesn't
+// exist.
 func ReadEnvDefault(key, _default string) string {
 	value, ok := os.LookupEnv(key)
 	if !ok {
@@ -24,12 +26,12 @@ func ReadEnvDefault(key, _default string) string {
 	return value
 }
 
-// ReadEnv reads an environment variable with a default value if it doesn't
+// ReadEnv reads an environment variable and raise an error if it doesn't
 // exist.
 func ReadEnv(key string) (string, error) {
 	value, ok := os.LookupEnv(key)
 	if !ok {
-		return "", errors.New("non-existed environment variables")
+		return "", errors.New("non-existed environment variables " + key)
 	}
 	return value, nil
 }
@@ -39,7 +41,7 @@ func ReadEnv(key string) (string, error) {
 func MustReadEnv(key string) string {
 	value, ok := os.LookupEnv(key)
 	if !ok {
-		log.Fatalln("non-existed key " + key)
+		log.Fatalln("non-existed environment variables " + key)
 	}
 
 	return value